change how selection works
Selection is now stored in a special container which has:
parallel arrays for selected ids and their groups.
A mapping between ent ids and indices into selected slice
A mapping between groups and the number of entities currently selected
which are part of that group.

This lets us store the order items are selected while also allowing
fast addition, removal, iteration of ids
and fast group lookups.

This commit also moves hidden sanitization of the selection into
Selection. This means ff an entity is unhidden which was selected when hidden, it
will no longer be selected.